Adjust fix for "Object name warning"
The fix that landed from #30 has a flaw if the same build report is viewed more than once using the Window option.
Because the timestamp is the same and the file already exists it was leaving an extra file "Library/LastBuild.buildreport" in the import folder.

Fix it by checking if the file exists and just select the existing one if it is already present.

Also add some comments.