AssetBundle TestProject and improved UI
Hide two tabs that are never displayed for AssetBundle builds
Put in a temporary workaround for absense of summary.buildType in older versions of Unity

Add a menu item for performing an AssetBundle build to the TestProject.  It demonstrates new recommended API and flags for 2022.3 and beyond but still works for 2021 case with a few workarounds.