v0.4.2 Add content summary tab (#40)
Add a "Content Summary" tab to the Build Report Inspector
This offers a simpler overview of the contents than the Source Asset tab.
The calculation happens in its own source file so its not intermingled with the UI for showing the content summary.

The display isn't beautiful as I'm not a UI programmer but hopefully it conveys the data clearly enough

Also:

Make a fix for issue 21. I couldn't reproduce, it may only be relevant for older versions of unity. But based on the callstack it was pretty clear how to avoid the exception.
Fix bug in my previous PR.