Expose SourceAsset calculation limit, and offer "export to CSV" capability
For large results the SourceAsset data can be best processed using a spreadsheet (with pivot tables) or a database.
Expose the maximum row limit in the UI, this controls both the CVS export and the UI display.

Although it takes an extra click to see the data, the extra step makes it possible to export a very large build, without actually displaying it in the build report inspector.