0.4.3 Source Asset Tab Improvements (#42)
Add ability to export SourceAsset data to CVS
Expose the maximum SourceAsset row count in the UI to remove hardcoded internal limit.

Added documentation with ideas for how to use the cvs data, including a demo script for exporting into SQLite and then performing a query on it.

Also:
Cleanout some extra files (.yamato directory)
Add a namespace for the ContentSummary.cs file, for consistency
Update the TestProject menus to offer ability to build a Player with compressed content