Test project: Add some AssetBundle Examples
Add example demonstrating different layouts for images in a simple scenario (this can lead to duplicated assets)

Incorporate an incremental build analysis helper that reporting what changes with performing an AssetBundle build.
This isn't directly related to BuildReportInspector because it looks at the build output itself.
It was originally published here:
https://discussions.unity.com/t/about-incremental-build-and-asset-bundle-hashes but potentially a little lost there.