Content Analysis - Avoid exception for built-in objects that have invalid filenames
Thanks to Sproilex for reporting and suggesting a fix!

This fix is based on pr #46 combined with the exisitng approach GetDisplayNameForFile() in BuildReportInspector.cs.

Instead of running a regex to sanitize each string we
catch exception in the rare case it is an invalid path name.
Its normal for there to be no extension on some of the paths that are processed.

I tested this by using the debugger to set one of the paths to example invalid path name
and make sure in the debugger and UI that Source Assets tab works ok, including the export to csv.